backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic[email]
passlib==1.7.4
bcrypt==3.2.2
python-jose[cryptography]
python-multipart
tenacity
alembic
pydantic-settings
